Added first version of a natural tile class Fixed issue where the functions for strings and files was declared in the header which caused issues when it is included in multiple files.
