+++ Opal Prospect/Opal Prospect.vcxproj
@@ -156,6 +156,8 @@
   <ItemGroup>
     <ClCompile Include="Cameras\Camera.cpp" />
     <ClCompile Include="Cameras\FlightCamera.cpp" />
+    <ClCompile Include="DwarfFortress\NaturalTile.cpp" />
+    <ClCompile Include="DwarfFortress\natural_tiles.cpp" />
     <ClCompile Include="FilePath.cpp" />
     <ClCompile Include="MainLoop.cpp" />
     <ClCompile Include="Opal Prospect.cpp" />
@@ -226,6 +228,7 @@
     <ClInclude Include="Constants\axis.hpp" />
     <ClInclude Include="Containers\ObjectPool.hpp" />
     <ClInclude Include="Containers\SortedObjectPool.hpp" />
+    <ClInclude Include="DwarfFortress\NaturalTile.hpp" />
     <ClInclude Include="DwarfFortress\natural_tiles.hpp" />
     <ClInclude Include="DwarfFortress\tile_types.hpp" />
     <ClInclude Include="FilePath.hpp" />
